requirements_pipeline.txt
# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0

# Configuration Management
pyyaml>=5.4.0

# Model Serialization
joblib>=1.1.0

# Experiment Tracking (optional but recommended)
mlflow>=1.20.0

# DVC Pipeline Orchestration
dvc>=2.0.0

# Utilities
python-dateutil>=2.8.0
